move all keen in a keenService
