+++ src/index.html
@@ -49,10 +49,10 @@
                     </div>
                 </div>
                 <div class="step1 hidden">
-                    <h2>Avez-vous votre clé secrette ?</h2>
+                    <h2>Avez-vous votre clé secrète ?</h2>
                     <p>La confrérie accueille actuellement 50 nouveaux membres, vous devez être invité par un membre pour pouvoir joindre.</p>
                     <div class="form-group">
-                        <label for="access-key">Clé d'accès secrette</label>
+                        <label for="access-key">Clé d'accès secrète</label>
                         <input type="text" id="access-key" class="form-control">
                     </div>
                     <button id="submit-access-key"
